Fix tokudb

+++ mock.sh
@@ -1,5 +1,19 @@
 #!/bin/bash
-set +e
+set -e
+
+if test -f /sys/kernel/mm/transparent_hugepage/enabled; then
+   echo "Need root permissions to disable /sys/kernel/mm/transparent_hugepage"
+   echo "If this fails, please run this script with sudo"
+   read -p "Press enter to continue or CTRL+C to cancel"
+   echo never > /sys/kernel/mm/transparent_hugepage/enabled
+fi
+if test -f /sys/kernel/mm/transparent_hugepage/defrag; then
+   echo "Need root permissions to disable /sys/kernel/mm/transparent_hugepage/defrag"
+   echo "If this fails, please run this script with sudo"
+   read -p "Press enter to continue or CTRL+C to cancel"
+   echo never > /sys/kernel/mm/transparent_hugepage/defrag
+fi
+
 # Remove any currently running containers
 docker stop mysql teamspeak mock-esi redis nsqlookup nsqadmin nsqd | xargs docker rm
 
@@ -19,6 +33,8 @@ docker run --net=host --name=redis -d -p 127.0.0.1:6379:6379 -h evedata.sql redi
 docker run --net=host --name=nsqlookup -d -p 127.0.0.1:4160:4160 -p 4161:4161 -h nsqlookupd1.nsq nsqio/nsq /nsqlookupd
 docker run --net=host --name=nsqadmin -d -p 127.0.0.1:4171:4171 -h nsqadmin.nsq nsqio/nsq /nsqadmin --lookupd-http-address=127.0.0.1:4161
 docker run --net=host --name=nsqd -d -p 127.0.0.1:4151:4151 -p 4150:4150 -h localhost nsqio/nsq /nsqd --lookupd-tcp-address=127.0.0.1:4160 -max-msg-size=8388608
+
+set +e
 
 # Get the admin token for the TS server
 until [ `docker inspect -f "{{.State.Status}}" teamspeak | grep -c running` -eq 1 ]
@@ -41,4 +57,4 @@ echo Percona Ready
 echo "create database eve; create database evedata; set sql_mode='STRICT_TRANS_TABLES,NO_ZERO_IN_DATE,NO_ZERO_DATE,ERROR_FOR_DIVISION_BY_ZERO';" | docker exec -i mysql /bin/bash -c 'mysql -uroot'
 cat ./services/vanguard/sql/evedata.sql | docker exec -i mysql /bin/bash -c 'mysql -uroot -Devedata'
 unzip -p ./services/vanguard/sql/eve.zip | docker exec -i mysql /bin/bash -c 'mysql -uroot -Deve'
-set -e
+
